pyproject.toml
[project]
name = "minecraft-avatar-generator"
version = "0.1.0"
description = "A command-line utility to generate 2D and 3D isometric renders of Minecraft skins from a player's username, UUID, or skin hash."
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "soisyhpe", email = "soisyhpe@gmail.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Topic :: Multimedia :: Graphics :: 3D Rendering",
    "Topic :: Utilities",
]
dependencies = [
    "pillow>=12.0.0",
    "numpy>=1.26.0",
    "attrs>=23.0.0",
]

[project.scripts]
mc-avatar = "cli:main"

[project.optional-dependencies]
dev = [
    "build",
    "twine",
]

[tool.setuptools]
py-modules = ["cli", "renderer"]
